chore: simplify namings for zustand
