+++ packages/react/src/components/console/ConsoleState.ts
@@ -4,8 +4,8 @@
  * MIT License
  */
 
-/* State */
-
+import { createStore } from 'zustand/vanilla';
+import { useStore } from 'zustand';
 import ConsoleAdapter from './ConsoleAdapter';
 
 export type IConsole = number;
@@ -14,3 +14,25 @@ export interface IConsoleState {
   outputs: IConsole;
   adapter?: ConsoleAdapter;
 }
+
+export type ConsoleState = IConsoleState & {
+  execute: (dark: boolean) => void;
+  setOutputs: (outputs: number) => void;
+  setAdapter: (adapter?: ConsoleAdapter) => void;
+};
+
+export const consoleStore = createStore<ConsoleState>((set, get) => ({
+  outputs: 0,
+  adapter: undefined,
+  execute: () => {},
+  setOutputs: (outputs: number) => set((state: ConsoleState) => ({ outputs })),
+  setAdapter: (adapter?: ConsoleAdapter) => set((state: ConsoleState) => ({ adapter })),
+}));
+
+export function useConsoleStore(): ConsoleState;
+export function useConsoleStore<T>(selector: (state: ConsoleState) => T): T;
+export function useConsoleStore<T>(selector?: (state: ConsoleState) => T) {
+  return useStore(consoleStore, selector!);
+}
+
+export default useConsoleStore;
