+++ packages/react/src/components/notebook/cell/sidebar/CellSidebarRun.tsx
@@ -9,7 +9,7 @@ import { PanelLayout } from '@lumino/widgets';
 import { Box, Button } from '@primer/react';
 import { PlayIcon } from '@primer/octicons-react';
 import { CellSidebarProps } from './CellSidebarWidget';
-import useNotebookStore from '../../NotebookZustand';
+import useNotebookStore from '../../NotebookState';
 
 import { DATALAYER_CELL_HEADER_CLASS } from './CellSidebarWidget';
