+++ packages/react/src/examples/NotebookExternalContent.tsx
@@ -12,7 +12,7 @@ import { INotebookContent } from '@jupyterlab/nbformat';
 import Jupyter from '../jupyter/Jupyter';
 import Notebook from '../components/notebook/Notebook';
 import CellSidebar from '../components/notebook/cell/sidebar/CellSidebar';
-import { notebookStore } from '../components/notebook/NotebookZustand';
+import { notebookStore } from '../components/notebook/NotebookState';
 
 import notebook from './notebooks/NotebookExample1.ipynb.json';
