+++ packages/react/src/index.ts
@@ -43,7 +43,6 @@ export * from './components/button/Button';
 export * from './components/cell/Cell';
 export * from './components/cell/CellAdapter';
 export * from './components/cell/CellState';
-export * from './components/cell/CellZustand';
 
 // CodeMirror Editor.
 // @todo CodeMirrorEditor breaks the JupyterLab extension loading.
@@ -56,7 +55,6 @@ export * from './components/commands/Commands';
 // Console.
 export * from './components/console/Console';
 export * from './components/console/ConsoleState';
-export * from './components/console/ConsoleZustand';
 
 // Dialog.
 export * from './components/dialog/Dialog';
@@ -92,7 +90,6 @@ export * from './components/lumino/ReactPortalWidget';
 export * from './components/notebook/Notebook';
 export * from './components/notebook/NotebookAdapter';
 export * from './components/notebook/NotebookState';
-export * from './components/notebook/NotebookZustand';
 export * from './components/notebook/content/JupyterReactContentFactory';
 export * from './components/notebook/cell/metadata/CellMetadataEditor';
 export * from './components/notebook/cell/metadata/NbGraderCells';
@@ -107,7 +104,6 @@ export * from './components/notebook/cell/sidebar/CellSidebarRun';
 // export * from './components/output/Output';
 export * from './components/output/OutputAdapter';
 export * from './components/output/OutputState';
-export * from './components/output/OutputZustand';
 export * from './components/output/OutputIPyWidgets';
 
 // Settings.
@@ -116,7 +112,6 @@ export * from './components/settings/Settings';
 // Terminal.
 export * from './components/terminal/Terminal';
 export * from './components/terminal/TerminalState';
-export * from './components/terminal/TerminalZustand';
 
 // TextInput
 export * from './components/textinput/TextInput';
