+++ packages/react/src/state/State.ts
@@ -10,11 +10,11 @@ import { useStore } from 'zustand';
 import { ServiceManager } from '@jupyterlab/services';
 import type { IDatalayerConfig } from './IState';
 import { IJupyterConfig, loadJupyterConfig } from '../jupyter/JupyterConfig';
-import useCellStore from '../components/cell/CellZustand';
-import useConsoleStore from '../components/console/ConsoleZustand';
-import useNotebookStore from '../components/notebook/NotebookZustand';
-import useOutputStore from '../components/output/OutputZustand';
-import useTerminalStore from '../components/terminal/TerminalZustand';
+import useCellStore from '../components/cell/CellState';
+import useConsoleStore from '../components/console/ConsoleState';
+import useNotebookStore from '../components/notebook/NotebookState';
+import useOutputStore from '../components/output/OutputState';
+import useTerminalStore from '../components/terminal/TerminalState';
 import { createLiteServer } from '../jupyter/lite/LiteServer';
 import { getJupyterServerHttpUrl } from '../jupyter/JupyterConfig';
 import { ensureJupyterAuth, createServerSettings, JupyterContextProps } from '../jupyter/JupyterContext';
